bug 2091: Define equals/hashCode on resource types
Resource will use a definition of equals that 
includes the content of the resource and its 
location. Location being loosely defined such that 
it can include the bundle location if installed or 
the repository location if in a repository. The 
other types will have equals defined based upon 
the members of the interface.

The newer Wiring type will not define equals/hashCode
and will default to identity.