bug 2475: rfc 185: Remove Serializable from DTO
In preparation to replace MapDTO with support for
List, Set and Map, Serializable is removed since
we cannot support serializing the subtypes from
the base DTO class. Each class in a hierarchy may
only control the serialization of the fields it
defines. So the DTO base class cannot handle the
serialization of fields defined by a subclass.
When the subclasses do not implement writeObject,
the VM will call defaultWriteObject for the
subclass. This will then serialize the concrete
forms (e.g. MyFancyList) of Lists, Sets and Maps
which we do not want. We also do not want to
serialization logic in every DTO subclass.

An option to be considered later would be to add
support for Externalizable. This is a subinterface
of Serializable which adds public writeExternal
and readExternal methods to DTO. These methods
would then have complete control over the
serialized form at the expense of public methods
which can be called by anyone.