ct: Fix invalid escaping of double quotes in a double quotes string
The backslash is removed by the properties file parsing done when Bnd
parses the bnd file. So then nested double quotes are no longer escaped
which creates a problem when Bnd 4.2 parses the improperly escaped
quotes in the string.

We change to use single quotes around the string to avoid the need to
properly escape (`\\"`) the double quotes in the string.